+++ src/models/user.js
@@ -9,6 +9,10 @@ const userSchema = new Schema({
     password: {
         type: String,
         required: true
+    },
+    email: {
+        type: String,
+        required: true
     }
 });
